third_party/WebKit/LayoutTests/fast/dom/shadow/hostcontext-first-child-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test that :host-context(:first-child) is re-evaluated when :first-child changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(first.firstChild, null).backgroundColor is red
PASS getComputedStyle(first.firstChild, null).backgroundColor is not red
PASS successfullyParsed is true

TEST COMPLETE
